mb/lenovo/sklkbl: Use spd_tools infrastructure for SPD binaries
For the t470s variant, this patch replaces the vendor SPD binaries for
soldered-on RAM with ones generated from spd_tools. This is in
preparation for adding variants with more complex onboard RAM
configurations.

This patch has been successfully validated on hardware (Thinkpad T470s
20JT-S16E00 with 4GB soldered-on RAM and unpopulated DIMM slot).

Depends-On: Ied92619130feaa160d01f75bc38230ab6a024ace
Reviewed-on: https://review.coreboot.org/c/coreboot/+/90027
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/mainboard/lenovo/sklkbl_thinkpad/variant.h
@@ -5,7 +5,7 @@
 
 void variant_config_gpios(void);
 
-int variant_memory_sku(void);
+uint8_t variant_memory_sku(void);
 
 void ssdt_add_dgpu(const struct device *dev);
